+++ public/javascripts/performanceReport.js
@@ -1,3 +1,4 @@
+let rateBTCUSD;
 
 const returnRateBTCTemplate = function returnCurrencyTemplate(obj) {
   let rate;
@@ -91,7 +92,7 @@ let performanceReportTableConfig = {
     { id: 'profitUSD', header:[null, {text: 'Worth USD', css: 'table-header-center' }], autowidth: true, autoheight: true, adjust: true, tooltip: tooltip, template: returnProfitUSDTemplate, cssFormat: alignRight },
     { id: 'wmr', header: [{text: '<span title="Weighted average rate of active loans">Weighted average rate<sup><i class="webix_icon fa-question-circle-o" style="font-size: 95%"></i></sup></span>', colspan: 2, css: 'table-header-center' },{ text: 'Daily', css: 'table-header-center' }], autowidth: false, adjust: true, sort: "int", tooltip: tooltip, template: returnWmrTemplate, cssFormat: alignRight, tooltip: 'Weighted average rate of active loans' },
     { id: 'ewmr', header: [null, {text: '<span title="Daily rate minus Poloniex commision">effective<sup><i class="webix_icon fa-question-circle-o" style="font-size: 95%"></i></sup></span>', css: 'table-header-center' }], autowidth: false, adjust: true, sort: "int", tooltip: tooltip, template: returnEwmrTemplate, cssFormat: alignRight, tooltip: 'daily rate minus Poloniex commision' },
-    { id: 'apy', header:[{ text: '<span title="Annual Percentage Yield">APY<sup><i class="webix_icon fa-question-circle-o" style="font-size: 95%"></i></sup></span>', css: 'table-header-center' }], autowidth: true, adjust: true, tooltip: tooltip, template: returnAPYTemplate, cssFormat: alignRight, tooltip: 'Annual Percentage Yield' },
+    { id: 'apy', header:[{ text: '<span title="Annual Percentage Yield\ntaking into account the effective daily rate and compound interest">APY<sup><i class="webix_icon fa-question-circle-o" style="font-size: 95%"></i></sup></span>', css: 'table-header-center' }], autowidth: true, adjust: true, tooltip: tooltip, template: returnAPYTemplate, cssFormat: alignRight, tooltip: 'Annual Percentage Yield\ntaking into account the effective daily rate and compound interest' },
   ],
   fixedRowHeight:false,  rowLineHeight:25, rowHeight:25,
 /*
@@ -112,6 +113,124 @@ let performanceReportTableConfig = {
   tooltip: true,
 };
 
+let btcAddress = '16oQxZYyFbS5Anju7EUAxTEUq7hBm3qfyu';
+let btcDonationsReceived = 0;
+let btcDonationsCount = 0;
+let usdDonationsReceived = 0;
+
+let donationsConfig = {
+  autoheight: true, borderless: true, type: 'clean',
+  rows: [
+    { template: 'poloLender is a free service<br> If it helped you in any way, if you would to contribute to the project and see it continue as a free service, <br>You can <b>make a donation</b> at BTC address below', autoheight: true,  },
+    {
+      cols: [
+        { view: 'text', width: inputTextWidth * 1.5, readonly: true, value: btcAddress, inputAlign: 'center',},
+        {
+          view: 'button',
+          id: `donationsDonateButton`,
+          tooltip: 'Donate bitcoin',
+          width: 60,
+          type: "htmlbutton", label:"<span class='webix_icon fa-bitcoin'></span>",
+          click: function () {
+            window.open(`bitcoin:${btcAddress}?amount=0.005`)
+          }
+        },
+        {
+          view: 'button',
+          id: `donationsCopyButton`,
+          tooltip: 'Copy address to clipboard',
+          width: 60,
+          type: "htmlbutton", label:"<span class='webix_icon fa-copy'></span>",
+          click: function () {
+            let text = btcAddress;
+            if (window.clipboardData && window.clipboardData.setData) {
+              // IE specific code path to prevent textarea being shown while dialog is visible.
+              return clipboardData.setData("Text", text);
+
+            } else if (document.queryCommandSupported && document.queryCommandSupported("copy")) {
+              let textarea = document.createElement("textarea");
+              textarea.textContent = text;
+              textarea.style.position = "fixed";  // Prevent scrolling to bottom of page in MS Edge.
+              document.body.appendChild(textarea);
+              textarea.select();
+              try {
+                document.execCommand("copy");  // Security exception may be thrown by some browsers.
+                webix.message('BTC address copied to clipboard');
+                return;
+              } catch (ex) {
+                console.warn("Copy to clipboard failed.", ex);
+                return false;
+              } finally {
+                document.body.removeChild(textarea);
+              }
+            }
+          }
+        },
+        {
+          view: 'button',
+          id: `donationsScanButton`,
+          tooltip: 'Scan QR Code',
+          width: 60,
+          type: "htmlbutton", label:"<span class='webix_icon fa-qrcode'></span>",
+          click: function () {
+            authUi = webix.ui({
+              view: 'window',
+              head: 'poloLender Pro donations',
+              modal: true,
+              position: 'center',
+              autoheight: true,
+              width: 450,
+              borderless: true, type: 'clean',
+              body: {
+                rows: [
+                  {gravity: 0.2},
+                  {
+                    cols: [
+                      {},
+                      {template: `<img src="http://i.imgur.com/cIfRJuU.png" alt="${btcAddress}">`, height: 160},
+                      {},
+                    ],
+                  },
+                  {
+                    cols: [
+                      {},
+                      { view: 'text', width: inputTextWidth * 1.5, readonly: true, value: btcAddress, inputAlign: 'center',},
+                      {},
+                    ],
+                  },
+                  {gravity: 0.2},
+                  {
+                    cols :[
+                      {},
+                      { view: 'button', width: buttonWidth, value: 'Close', type: 'danger',
+                        click: function (elementId, event) {
+                          authUi.hide();
+                        }
+                      },
+                      {gravity: 0.05},
+                    ]
+                  }
+                ],
+              },
+              move: true,
+            });
+            authUi.show();
+          }
+        },
+        {},
+      ]
+    },
+    { id: 'dotationsSoFar', template: function (obj) {
+      let msg  = `<br>${btcDonationsCount} donations since July 1, 2017: BTC ${btcDonationsReceived} (USD ${Math.round(btcDonationsReceived * rateBTCUSD) || 0})`;
+      msg += '<br>If you have made a donation please also feel free to <a href="mailto:dutu@protonmail.com">drop me a message</a>.';
+      msg += '<br>We continue adding features and developing new software, contributors will have priority access to free and premium features/services.';
+      return msg;
+    }},
+    {},
+  ]
+
+};
+
 let performanceReportView = {
   id: 'performanceReport',
   borderless: true,
@@ -124,6 +243,9 @@ let performanceReportView = {
       rows: [
         { view:"template", template:"Performance report", type:"section" },
         performanceReportTableConfig,
+        { gravity: 0.2 },
+        { view:"template", template:"Donations", type:"section" },
+        donationsConfig,
         { gravity: 1 },
       ]
     },
@@ -131,11 +253,20 @@ let performanceReportView = {
   ],
 };
 
+
 let performanceReportTable = [];
-let updatePerformanceReport = function updatePerformanceReport(data) {
+let performanceReportData = {};
+
+let refreshPerformanceView = function refreshPerformanceView() {
+  let visibleView = $$('contentTabview').getMultiview().getValue();
+  if (visibleView !== 'performanceView') {
+    return;
+  }
+
+  $$('dotationsSoFar').refresh();
   let runningDays = parseFloat(moment().diff(poloLenderAppStatusData.runningSince, "minutes", true).toString()) / 60 / 24;
   lendingCurrencies.slice().reverse().forEach((currency) => {
-    let performanceData = data[currency];
+    let performanceData = performanceReportData[currency];
     let existingCurrencyDataIndex = _.findIndex(performanceReportTable, { currency: currency });
 
     if (existingCurrencyDataIndex > -1 && !performanceData) {
@@ -172,6 +303,7 @@ let updatePerformanceReport = function updatePerformanceReport(data) {
     newRow.rateBTCChange = existingCurrencyData && (Math.sign(newRow.rateBTC - existingCurrencyData.rateBTC) || existingCurrencyData.rateBTCChange) || 0;
     newRow.rateBTCUSDChange = existingCurrencyData && (Math.sign(newRow.rateBTCUSD - existingCurrencyData.rateBTCUSD) || existingCurrencyData.rateBTCUSDChange) || 0;
     newRow.wmrChange = existingCurrencyData && (Math.sign(newRow.wmr - existingCurrencyData.wmr) || existingCurrencyData.wmrChange) || 0;
+    rateBTCUSD = parseFloat(performanceData.rateBTCUSD) || 0;
     if (existingCurrencyDataIndex === -1) {
       performanceReportTable.push(newRow);
     } else {
@@ -189,3 +321,19 @@ let updatePerformanceReport = function updatePerformanceReport(data) {
     performanceReportTableUi.adjustColumn(row, 'data');
   });
 };
+
+let startUpdateDonations = function startUpdateDonations() {
+  webix.ajax(`https://blockexplorer.com/api/addr/${btcAddress}`, function (text,data) {
+    let jsonData = JSON.parse(text);
+    btcDonationsReceived = jsonData.totalReceived || 0;
+    btcDonationsCount = jsonData.txApperances || 0;
+    $$('dotationsSoFar').refresh();
+    setTimeout(startUpdateDonations, 60 * 1000);
+  });
+};
+
+let updatePerformanceReport = function updatePerformanceReport(data) {
+  performanceReportData = data;
+  refreshPerformanceView();
+};
+