+++ public/javascripts/status.js
@@ -151,6 +151,7 @@ let lendingEngineStatusConfig = {
             }
 
             config.isTradingEnabled = !config.isTradingEnabled;
+            config.status.lendingEngineStopReason = 'Manually stopped';
             showProcessingDataMessage();
             socket.emit('updateConfig', config);
           },
@@ -268,6 +269,12 @@ let updatePoloLenderAppStatus = function updatePoloLenderAppStatus() {
     restartedAt: status.restarted,
     message: clientMessage.message,
   };
+
+  let visibleView = $$('contentTabview').getMultiview().getValue();
+  if (visibleView !== 'statusView') {
+    return;
+  }
+
   let poloLenderApp_runningSinceUi = $$('poloLenderApp_runningSince');
   if (poloLenderApp_runningSinceUi) poloLenderApp_runningSinceUi.refresh();
   let poloLenderApp_versionUi = $$('poloLenderApp_version');
@@ -284,6 +291,12 @@ let updateLendingEngineStatus = function updateLendingEngineStatus() {
     lendingEngineStopTime: config.status && config.status.lendingEngineStopTime || '',
     lendingEngineStopReason: config.status && config.status.lendingEngineStopReason || '',
   };
+
+  let visibleView = $$('contentTabview').getMultiview().getValue();
+  if (visibleView !== 'statusView') {
+    return;
+  }
+
   $$('lendingEngineStatus_status').refresh();
   $$('lendingEngineStatus_apiActivity').refresh();
   let lendingEngineStartStopButtonUi = $$('lendingEngineStartStopButton');
@@ -302,12 +315,25 @@ let updateAdvisorEngineStatus = function updateAdvisorEngineStatus() {
     connection: status.lendingAdvisor && status.lendingAdvisor.connection || '',
     authentication: status.lendingAdvisor && status.lendingAdvisor.authentication || { status: 0, message: '' },
   };
+
+  let visibleView = $$('contentTabview').getMultiview().getValue();
+  if (visibleView !== 'statusView') {
+    return;
+  }
+
   $$('advisorEngine_server').refresh();
   $$('advisorEngine_connectionStatus').refresh();
 };
 
 let advisorInfoTable = [];
-let updateAdvisorInfo = function updateAdvisorInfo(advisorInfo) {
+let advisorInfo = {};
+
+let refreshAdvisorInfo = function refreshAdvisorInfo() {
+  let visibleView = $$('contentTabview').getMultiview().getValue();
+  if (visibleView !== 'statusView') {
+    return;
+  }
+
   _.forEach(advisorInfo, (value, key) => {
     if (key === 'time') {
       return;
@@ -341,11 +367,21 @@ let updateAdvisorInfo = function updateAdvisorInfo(advisorInfo) {
   advisorInfoTableUi.refreshColumns();
 };
 
+let updateAdvisorInfo = function updateAdvisorInfo(data) {
+  advisorInfo = data;
+  refreshAdvisorInfo();
+};
+
 let advisorInfoTableUi;
 let lendingEngineStatus_apiActivityUi = null;
 
 let startRefreshingStatus = function startRefreshingStatus() {
   setInterval(function refreshPoloLenderAppStatus() {
+    let visibleView = $$('contentTabview').getMultiview().getValue();
+    if (visibleView !== 'statusView') {
+      return;
+    }
+
     $$('poloLenderApp_runningSince').refresh();
     $$('poloLenderApp_restartedAt').refresh();
     $$('poloLenderApp_clientMessage').refresh();
@@ -353,6 +389,18 @@ let startRefreshingStatus = function startRefreshingStatus() {
 
   lendingEngineStatus_apiActivityUi = $$('lendingEngineStatus_apiActivity');
   setInterval(function refreshPoloLenderAppInfo() {
+    let visibleView = $$('contentTabview').getMultiview().getValue();
+    if (visibleView !== 'statusView') {
+      return;
+    }
+
     lendingEngineStatus_apiActivityUi.refresh();
-  }, 50);
+  }, 1000/8);
+};
+
+let refreshStatusView = function refreshStatusView() {
+  refreshAdvisorInfo();
+  updateAdvisorEngineStatus();
+  updateLendingEngineStatus();
+  updatePoloLenderAppStatus();
 };