+++ server/workers/poloLender/poloLender.js
@@ -16,6 +16,7 @@ import { migrateConfig } from './dbMigrate';
 import { io } from '../../httpServer';
 import { connectDb, getConfig, saveConfig } from './config';
 import { debugApiCallDuration, debugCycleDuration, debugTimer } from './debug';
+import { getLogTrailItems } from "./logtrail";
 
 debug('pololender');
 
@@ -79,6 +80,11 @@ export const PoloLender = function(name) {
   const isWriteAllowed = function isWriteAllowed(cliendId) {
     let authClient = _.find(authClients, { id: cliendId });
     return authClient && authClient.isReadWriteAllowed;
+  };
+
+  const isReadAllowed = function isReadAllowed(cliendId) {
+    let authClient = _.find(authClients, { id: cliendId });
+    return authClient && authClient.isReadAllowed;
   };
 
   const authClientsEmit = function authClientsEmit() {
@@ -204,9 +210,9 @@ export const PoloLender = function(name) {
     };
     saveConfig(config);
 
-    log.info(`Your read/only authorization token is: ${config.authToken.readOnly}`);
-    log.info(`Your read/write authorization token is: ${config.authToken.readWrite}`);
-    log.info(`Token expires on: ${config.authToken.tokenExpiresOn}`);
+    console.log(`Your read/only authorization token is: ${config.authToken.readOnly}`);
+    console.log(`Your read/write authorization token is: ${config.authToken.readWrite}`);
+    console.log(`Token expires on: ${config.authToken.tokenExpiresOn}`);
 
     let authClient = {
       id: clientId,
@@ -249,11 +255,26 @@ export const PoloLender = function(name) {
     });
   };
 
+  const onReturnLogtrailBuffer = function onReturnLogtrailBuffer(params) {
+    let clientId = this.id;
+    if (!isReadAllowed(clientId)) {
+      return;
+    }
+
+    getLogTrailItems(params, (errMessage, result) => {
+      let client = io.sockets.clients().connected[clientId];
+      if (client) {
+        client.emit('logtrailBuffer', errMessage, result);
+      }
+    });
+  };
+
   const onBrowserConnection = function onBrowserConnection(socket) {
     socket.on('authorization', onAuthorization);
     socket.on(`validateToken`, onValidateToken);
     socket.on(`generateNewToken`, onGenerateNewToken);
     socket.on('returnLendingHistory', onReturnLendingHistory);
+    socket.on('returnLogtrailBuffer', onReturnLogtrailBuffer);
     socket.on(`updateConfig`, onUpdateConfig);
 
   };
@@ -531,20 +552,19 @@ export const PoloLender = function(name) {
             let offerRate;
             let amountTrading;
 
-            offerRate = new Big(offer.rate);
-            if (offerRate.eq(advisorInfo[currency] && advisorInfo[currency].bestReturnRate || '0.05')){
+            offerRate = parseFloat(offer.rate);
+            let recommendedRate = parseFloat(advisorInfo[currency] && advisorInfo[currency].bestReturnRate || '0.05');
+            let minRate = parseFloat(new Big(config.offerMinRate[currency] || 0).div(100).toString());
+            recommendedRate = Math.max(minRate, recommendedRate);
+            if (offerRate === recommendedRate){
               // lend offers is on correct price
-              return cb(null);
-            }
-
-            if (!(config.offerMaxAmount[currency] === "")) {
-              // only if we are reserving any amount check if we are already trading more then offerMaxAmount
               amountTrading = new Big(depositFunds[currency]).minus(availableFunds[currency]);
-              if(amountTrading.gte(config.offerMaxAmount[currency] || 9999999)) {
+              if(amountTrading.lte(config.offerMaxAmount[currency] || 9999999)) {
                 // we are already trading higher then offerMaxAmount
                 return cb(null);
               }
             }
+
             if (process.env[self.me+"_NOTRADE"] === "true") {
               log.notice("cancelHighOffers: NO TRADE");
               return cb(null);
@@ -789,6 +809,7 @@ export const PoloLender = function(name) {
   };
 
   let currentApiKey = {};
+  let currentTelegramUserInfo = {};
   const execTrade = function execTrade() {
     async.series(
       {
@@ -801,6 +822,12 @@ export const PoloLender = function(name) {
             if (!_.isEqual(currentApiKey, config.apiKey)) {
               currentApiKey = config.apiKey;
               poloPrivate = new Poloniex(currentApiKey.key, currentApiKey.secret, { socketTimeout: 60000 });
+            }
+
+            if (currentTelegramUserInfo.telegramUserId !== config.telegramReports.telegramUserId || currentTelegramUserInfo.telegramToken !== config.telegramReports.telegramToken) {
+              currentTelegramUserInfo.telegramUserId = config.telegramReports.telegramUserId;
+              currentTelegramUserInfo.telegramToken = config.telegramReports.telegramToken;
+              logTg = addTelegramLogger(config.telegramReports && config.telegramReports.telegramToken, config.telegramReports && config.telegramReports.telegramUserId);
             }
 
             if (!config.isTradingEnabled && newConfig.isTradingEnabled) {
@@ -1033,9 +1060,9 @@ export const PoloLender = function(name) {
             saveConfig(config)
           }
 
-          log.info(`Your read/only authorization token is: ${config.authToken.readOnly}`);
-          log.info(`Your read/write authorization token is: ${config.authToken.readWrite}`);
-          log.info(`Token expires on: ${config.authToken.tokenExpiresOn}`);
+          console.log(`Your read/only authorization token is: ${config.authToken.readOnly}`);
+          console.log(`Your read/write authorization token is: ${config.authToken.readWrite}`);
+          console.log(`Token expires on: ${config.authToken.tokenExpiresOn}`);
 
           httpServerStart(config.port);
           currentApiKey = config.apiKey;